+++ src/libs/conduit/conduit_data_type.hpp
@@ -526,6 +526,8 @@ public:
 // Transforms
 //-----------------------------------------------------------------------------
     std::string         to_string(const std::string &protocol="json") const;
+    void                to_string_stream(std::ostream &os, 
+                                         const std::string &protocol="json") const;
 
     // NOTE(cyrush): The primary reason this function exists is to enable 
     // easier compatibility with debugging tools (e.g. totalview, gdb) that